# Use intervals instead of manual ping timeout calculation