As suggested in #5, I've rewritten ping interval to the `tokio::interval` instead of manual range calculation.

I've suggested replacing stream timeouts to the similar solution, because the `SystemTimer` is not monotonic one.

Benchmark before:
```
     Running benches/bench.rs (target/release/deps/bench-a5e9977bff1da307)

running 2 tests
test tests::bench_mux_send ... bench:      32,755.78 ns/iter (+/- 9,696.66) = 4001 MB/s
test tests::bench_tcp_send ... bench:      12,050.36 ns/iter (+/- 914.29) = 10877 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 65.31s
```

Benchmark after:
```
     Running benches/bench.rs (target/release/deps/bench-a5e9977bff1da307)

running 2 tests
test tests::bench_mux_send ... bench:      32,561.84 ns/iter (+/- 6,245.88) = 4025 MB/s
test tests::bench_tcp_send ... bench:      12,072.64 ns/iter (+/- 901.40) = 10857 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 67.56s
```